afsd: correct printf format mismatch in debugging printf
On platforms where size_t is unsigned int, the type of
cacheFiles * sizeof(AFSD_INO_T) is not an unsigned long as the format
string requires.  Casting cacheFiles to unsigned long ensures that the
result is at least unsigned long, although it will still be wrong if
any architecture makes size_t be long long.  Fixes build for FreeBSD
10 on i386.

Reviewed-on: http://gerrit.openafs.org/11384
Tested-by: BuildBot <buildbot@rampaginggeek.com>
(cherry picked from commit f02beb8d644ec2a52bf21737a040321905a39e20)

Reviewed-on: http://gerrit.openafs.org/11403
Tested-by: BuildBot <buildbot@rampaginggeek.com>

+++ src/afsd/afsd.c
@@ -2121,7 +2121,8 @@ afsd_run(void)
     memset(inode_for_V, '\0', (cacheFiles * sizeof(AFSD_INO_T)));
     if (afsd_debug)
 	printf("%s: %d inode_for_V entries at %p, %lu bytes\n", rn,
-	       cacheFiles, inode_for_V, (cacheFiles * sizeof(AFSD_INO_T)));
+	       cacheFiles, inode_for_V,
+	       (unsigned long)cacheFiles * sizeof(AFSD_INO_T));
 #endif
 
     if (!(cacheFlags & AFSCALL_INIT_MEMCACHE)) {
